Added controllers
seperated the logic for the database to be done via the controllers this follows best practices and makes code more modular and maintanable
